[image] move colorspace code into dedicated file
