+++ src/aliceVision/image/io.hpp
@@ -7,8 +7,10 @@
 
 #pragma once
 
-#include <aliceVision/image/Image.hpp>
-#include <aliceVision/image/pixelTypes.hpp>
+#include "Image.hpp"
+#include "pixelTypes.hpp"
+#include "colorspace.hpp"
+
 #include <aliceVision/types.hpp>
 
 #include <OpenImageIO/paramlist.h>
@@ -17,38 +19,12 @@
 
 #include <string>
 
-namespace oiio = OIIO;
 
 namespace aliceVision {
 
 class rgb;
 
 namespace image {
-
-/**
- * @brief Available image color space for pipeline input
- */
-enum class EImageColorSpace
-{
-  AUTO,
-  LINEAR,
-  SRGB,
-  ACES2065_1,
-  ACEScg,
-  LAB,
-  XYZ,
-  NO_CONVERSION
-};
-
-std::string EImageColorSpace_informations();
-EImageColorSpace EImageColorSpace_stringToEnum(const std::string& dataType);
-std::string EImageColorSpace_enumToString(const EImageColorSpace dataType);
-std::string EImageColorSpace_enumToOIIOString(const EImageColorSpace colorSpace);
-EImageColorSpace EImageColorSpace_OIIOstringToEnum(const std::string& colorspace);
-bool EImageColorSpace_isSupportedOIIOEnum(const EImageColorSpace& colorspace);
-bool EImageColorSpace_isSupportedOIIOstring(const std::string& colorspace);
-std::ostream& operator<<(std::ostream& os, EImageColorSpace dataType);
-std::istream& operator>>(std::istream& in, EImageColorSpace& dataType);
 
 struct OutputFileColorSpace
 {
